API redis health check
