--- modules/back-end/src/Infrastructure/Caches/Redis/IRedisClient.cs
@@ -1,8 +1,0 @@
-using StackExchange.Redis;
-
-namespace Infrastructure.Caches.Redis;
-
-public interface IRedisClient
-{
-    IDatabase GetDatabase();
-}
