+++ modules/back-end/src/Infrastructure/Caches/Redis/RedisPopulatingService.cs
@@ -18,7 +18,7 @@ public class RedisPopulatingService : ICachePopulatingService
     private readonly ILogger<RedisPopulatingService> _logger;
 
     public RedisPopulatingService(
-        IRedisClient redis,
+        IConnectionMultiplexer redis,
         ICacheService cache,
         IFeatureFlagService flagService,
         ISegmentService segmentService,
