+++ modules/back-end/src/Infrastructure/ConfigurationExtensions.cs
@@ -2,6 +2,7 @@
 using Infrastructure.MQ;
 using Infrastructure.Persistence;
 using Microsoft.Extensions.Configuration;
+using StackExchange.Redis;
 
 namespace Infrastructure;
 
@@ -30,4 +31,22 @@ public static class ConfigurationExtensions
         var provider = configuration.GetValue(CacheProvider.SectionName, CacheProvider.Redis)!;
         return provider;
     }
+
+    public static string GetRedisConnectionString(this IConfiguration configuration)
+    {
+        var connectionString = configuration["Redis:ConnectionString"];
+        var options = ConfigurationOptions.Parse(connectionString!);
+
+        // if we specified a password in the configuration, use it
+        var password = configuration["Redis:Password"];
+        if (!string.IsNullOrWhiteSpace(password))
+        {
+            options.Password = password;
+        }
+
+        return options.ToString(includePassword: true);
+    }
+
+    public static string GetPostgresConnectionString(this IConfiguration configuration)
+        => configuration["Postgres:ConnectionString"]!;
 }
