+++ modules/back-end/src/Infrastructure/ServiceCollectionExtensions.cs
@@ -7,6 +7,7 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Npgsql;
+using StackExchange.Redis;
 
 namespace Infrastructure;
 
@@ -23,7 +24,7 @@ public static class ServiceCollectionExtensions
 
     public static void TryAddRedis(this IServiceCollection services, IConfiguration configuration)
     {
-        if (services.Any(service => service.ServiceType == typeof(IRedisClient)))
+        if (services.Any(service => service.ServiceType == typeof(IConnectionMultiplexer)))
         {
             return;
         }
@@ -32,7 +33,8 @@ public static class ServiceCollectionExtensions
             .Bind(configuration.GetSection(RedisOptions.Redis))
             .ValidateDataAnnotations();
 
-        services.AddSingleton<IRedisClient, DefaultRedisClient>();
+        var multiplexer = ConnectionMultiplexer.Connect(configuration.GetRedisConnectionString());
+        services.AddSingleton<IConnectionMultiplexer>(multiplexer);
     }
 
     public static void TryAddMongoDb(this IServiceCollection services, IConfiguration configuration)
@@ -63,7 +65,7 @@ public static class ServiceCollectionExtensions
         // https://github.com/npgsql/efcore.pg/issues/1107
         // https://github.com/npgsql/efcore.pg/issues/3119
 
-        services.AddNpgsqlDataSource(configuration["Postgres:ConnectionString"]!, builder =>
+        services.AddNpgsqlDataSource(configuration.GetPostgresConnectionString(), builder =>
         {
             builder
                 .EnableDynamicJson()
